chore(client): refactor
