+++ client/src/components/Domain/Bisc/Suppliers/DeleteSupplierModal.tsx
@@ -2,15 +2,15 @@ import { t } from '@lingui/macro'
 import { useLingui } from '@lingui/react'
 import React from 'react'
 import { useHistory } from 'react-router-dom'
-import Button, { ButtonType } from '../../../../components/Core/Button/Button'
-import { NotificationsManager } from '../../../../components/Core/Feedback/Notifications/NotificationsManager'
-import { IconType } from '../../../../components/Core/Icon/IconType'
-import Column from '../../../../components/Core/Layout/Column/Column'
-import ModalView from '../../../../components/Core/Modal/ModalView'
-import SectionTitle from '../../../../components/Core/Text/SectionTitle'
-import Paragraph from '../../../../components/Core/Typography/Paragraph'
-import { AanbiedersDocument, useDeleteAanbiederMutation } from '../../../../generated/graphql'
-import { routes } from '../../../../routes/routes'
+import Button, { ButtonType } from 'components/Core/Button/Button'
+import { NotificationsManager } from 'components/Core/Feedback/Notifications/NotificationsManager'
+import { IconType } from 'components/Core/Icon/IconType'
+import Column from 'components/Core/Layout/Column/Column'
+import ModalView from 'components/Core/Modal/ModalView'
+import SectionTitle from 'components/Core/Text/SectionTitle'
+import Paragraph from 'components/Core/Typography/Paragraph'
+import { AanbiedersDocument, useDeleteAanbiederMutation } from 'generated/graphql'
+import { routes } from 'routes/routes'
 
 interface Props {
     onClose: () => void
