+++ client/src/views/Authorized/Supplier/BiscView/SupplierCreateView.tsx
@@ -2,25 +2,25 @@ import { t } from '@lingui/macro'
 import { useLingui } from '@lingui/react'
 import React from 'react'
 import { useHistory } from 'react-router-dom'
-import Headline from '../../../components/Chrome/Headline'
-import Actionbar from '../../../components/Core/Actionbar/Actionbar'
-import Breadcrumb from '../../../components/Core/Breadcrumb/Breadcrumb'
-import Breadcrumbs from '../../../components/Core/Breadcrumb/Breadcrumbs'
-import Button, { ButtonType } from '../../../components/Core/Button/Button'
-import { NotificationsManager } from '../../../components/Core/Feedback/Notifications/NotificationsManager'
-import Form from '../../../components/Core/Form/Form'
-import HorizontalRule from '../../../components/Core/HorizontalRule/HorizontalRule'
-import Row from '../../../components/Core/Layout/Row/Row'
-import Space from '../../../components/Core/Layout/Space/Space'
+import Headline from 'components/Chrome/Headline'
+import Actionbar from 'components/Core/Actionbar/Actionbar'
+import Breadcrumb from 'components/Core/Breadcrumb/Breadcrumb'
+import Breadcrumbs from 'components/Core/Breadcrumb/Breadcrumbs'
+import Button, { ButtonType } from 'components/Core/Button/Button'
+import { NotificationsManager } from 'components/Core/Feedback/Notifications/NotificationsManager'
+import Form from 'components/Core/Form/Form'
+import HorizontalRule from 'components/Core/HorizontalRule/HorizontalRule'
+import Row from 'components/Core/Layout/Row/Row'
+import Space from 'components/Core/Layout/Space/Space'
 import BranchInformationFieldset, {
     BranchInformationFieldsetFormModel,
-} from '../../../components/fieldsets/shared/BranchInformationFieldset'
+} from 'components/fieldsets/shared/BranchInformationFieldset'
 import ContactInformationFieldset, {
     ContactInformationFieldsetModel,
-} from '../../../components/fieldsets/shared/ContactInformationFieldset'
-import { AanbiedersDocument, useCreateAanbiederMutation } from '../../../generated/graphql'
-import { routes } from '../../../routes/routes'
-import { Forms } from '../../../utils/forms'
+} from 'components/fieldsets/shared/ContactInformationFieldset'
+import { AanbiedersDocument, useCreateAanbiederMutation } from 'generated/graphql'
+import { routes } from 'routes/routes'
+import { Forms } from 'utils/forms'
 
 interface FormModel extends BranchInformationFieldsetFormModel, ContactInformationFieldsetModel {}
