+++ client/src/views/Authorized/Supplier/BiscView/SupplierDetailView/Coworkers/CoworkerDetail/CoworkerDetailDataUpdateView.tsx
@@ -2,34 +2,30 @@ import { t } from '@lingui/macro'
 import { useLingui } from '@lingui/react'
 import React from 'react'
 import { useHistory, useParams } from 'react-router-dom'
-import Headline from '../../../../../../../components/Chrome/Headline'
-import Actionbar from '../../../../../../../components/Core/Actionbar/Actionbar'
-import Breadcrumb from '../../../../../../../components/Core/Breadcrumb/Breadcrumb'
-import Breadcrumbs from '../../../../../../../components/Core/Breadcrumb/Breadcrumbs'
-import Button, { ButtonType } from '../../../../../../../components/Core/Button/Button'
-import ErrorBlock from '../../../../../../../components/Core/Feedback/Error/ErrorBlock'
-import { NotificationsManager } from '../../../../../../../components/Core/Feedback/Notifications/NotificationsManager'
-import Spinner, { Animation } from '../../../../../../../components/Core/Feedback/Spinner/Spinner'
-import Form from '../../../../../../../components/Core/Form/Form'
-import HorizontalRule from '../../../../../../../components/Core/HorizontalRule/HorizontalRule'
-import Center from '../../../../../../../components/Core/Layout/Center/Center'
-import Row from '../../../../../../../components/Core/Layout/Row/Row'
-import Space from '../../../../../../../components/Core/Layout/Space/Space'
+import Headline from 'components/Chrome/Headline'
+import Actionbar from 'components/Core/Actionbar/Actionbar'
+import Breadcrumb from 'components/Core/Breadcrumb/Breadcrumb'
+import Breadcrumbs from 'components/Core/Breadcrumb/Breadcrumbs'
+import Button, { ButtonType } from 'components/Core/Button/Button'
+import ErrorBlock from 'components/Core/Feedback/Error/ErrorBlock'
+import { NotificationsManager } from 'components/Core/Feedback/Notifications/NotificationsManager'
+import Spinner, { Animation } from 'components/Core/Feedback/Spinner/Spinner'
+import Form from 'components/Core/Form/Form'
+import HorizontalRule from 'components/Core/HorizontalRule/HorizontalRule'
+import Center from 'components/Core/Layout/Center/Center'
+import Row from 'components/Core/Layout/Row/Row'
+import Space from 'components/Core/Layout/Space/Space'
 import AccountInformationFieldset, {
     AccountInformationFieldsetModel,
-} from '../../../../../../../components/fieldsets/shared/AccountInformationFieldset'
-import AvailabillityFieldset, {
-    AvailabillityFieldsetModel,
-} from '../../../../../../../components/fieldsets/shared/AvailabillityFieldset'
-import InformationFieldset, {
-    InformationFieldsetModel,
-} from '../../../../../../../components/fieldsets/shared/InformationFieldset'
-import { useMockQuery } from '../../../../../../../components/hooks/useMockQuery'
-import { useMockMutation } from '../../../../../../../hooks/UseMockMutation'
-import { routes } from '../../../../../../../routes/routes'
-import { SupplierDetailCoworkersParams } from '../../../../../../../routes/supplier/types'
-import { Forms } from '../../../../../../../utils/forms'
-import { coworkerDetailMock, CoworkerDetailResponseMock, coworkersCreateMock } from '../../mocks/coworkers'
+} from 'components/fieldsets/shared/AccountInformationFieldset'
+import AvailabillityFieldset, { AvailabillityFieldsetModel } from 'components/fieldsets/shared/AvailabillityFieldset'
+import InformationFieldset, { InformationFieldsetModel } from 'components/fieldsets/shared/InformationFieldset'
+import { useMockQuery } from 'components/hooks/useMockQuery'
+import { useMockMutation } from 'hooks/UseMockMutation'
+import { routes } from 'routes/routes'
+import { SupplierDetailCoworkersParams } from 'routes/supplier/types'
+import { Forms } from 'utils/forms'
+import { coworkerDetailMock, CoworkerDetailResponseMock, coworkersCreateMock } from '../mocks/coworkers'
 
 interface Props {}
