+++ client/src/views/Authorized/Supplier/BiscView/SupplierDetailView/Coworkers/CoworkerDetail/CoworkerDetailView.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { Redirect, Route, Switch } from 'react-router-dom'
-import { routes } from '../../../../../../routes/routes'
-import { NotFoundView } from '../../../../../Generic/NotFoundView'
-import CoworkerDetailDataUpdateView from './CoworkerDetailData/CoworkerDetailDataUpdateView'
-import CoworkerDetailDataView from './CoworkerDetailData/CoworkerDetailDataView'
-import CoworkerDetailDocumentsView from './CoworkerDetailDocuments/CoworkerDetailDocumentsView'
+import { routes } from 'routes/routes'
+import { NotFoundView } from 'views/Generic/NotFoundView'
+import CoworkerDetailDataUpdateView from './CoworkerDetailDataUpdateView'
+import CoworkerDetailDataView from './CoworkerDetailDataView'
+import CoworkerDetailDocumentsView from './CoworkerDetailDocumentsView'
 
 interface Props {}
