+++ client/src/views/Authorized/Supplier/BiscView/SupplierDetailView/Coworkers/CoworkersCreateView.tsx
@@ -2,30 +2,30 @@ import { t } from '@lingui/macro'
 import { useLingui } from '@lingui/react'
 import React, { useState } from 'react'
 import { useHistory, useParams } from 'react-router-dom'
-import Headline from '../../../../../components/Chrome/Headline'
-import Actionbar from '../../../../../components/Core/Actionbar/Actionbar'
-import Breadcrumb from '../../../../../components/Core/Breadcrumb/Breadcrumb'
-import Breadcrumbs from '../../../../../components/Core/Breadcrumb/Breadcrumbs'
-import Button, { ButtonType } from '../../../../../components/Core/Button/Button'
-import { NotificationsManager } from '../../../../../components/Core/Feedback/Notifications/NotificationsManager'
-import Form from '../../../../../components/Core/Form/Form'
-import HorizontalRule from '../../../../../components/Core/HorizontalRule/HorizontalRule'
-import { IconType } from '../../../../../components/Core/Icon/IconType'
-import Row from '../../../../../components/Core/Layout/Row/Row'
-import Space from '../../../../../components/Core/Layout/Space/Space'
-import SectionTitle from '../../../../../components/Core/Text/SectionTitle'
-import AccountInformationFieldset from '../../../../../components/fieldsets/shared/AccountInformationFieldset'
-import AvailabillityFieldset from '../../../../../components/fieldsets/shared/AvailabillityFieldset'
-import ContactPersonInformationFieldset from '../../../../../components/fieldsets/shared/ContactPersonInformationFieldset'
-import ContactInformationFieldset from '../../../../../components/fieldsets/shared/ContactPersonInformationFieldset'
-import CourseInformationFieldset from '../../../../../components/fieldsets/shared/CourseInformationFieldset'
-import EducationInformationFieldset from '../../../../../components/fieldsets/shared/EducationInformationFieldset'
-import InformationFieldset from '../../../../../components/fieldsets/shared/InformationFieldset'
-import PersonInformationFieldset, { Roles } from '../../../../../components/fieldsets/shared/PersonInformationFieldset'
-import { useMockMutation } from '../../../../../hooks/UseMockMutation'
-import { routes } from '../../../../../routes/routes'
-import { SupplierDetailParams } from '../../../../../routes/supplier/types'
-import { Forms } from '../../../../../utils/forms'
+import Headline from 'components/Chrome/Headline'
+import Actionbar from 'components/Core/Actionbar/Actionbar'
+import Breadcrumb from 'components/Core/Breadcrumb/Breadcrumb'
+import Breadcrumbs from 'components/Core/Breadcrumb/Breadcrumbs'
+import Button, { ButtonType } from 'components/Core/Button/Button'
+import { NotificationsManager } from 'components/Core/Feedback/Notifications/NotificationsManager'
+import Form from 'components/Core/Form/Form'
+import HorizontalRule from 'components/Core/HorizontalRule/HorizontalRule'
+import { IconType } from 'components/Core/Icon/IconType'
+import Row from 'components/Core/Layout/Row/Row'
+import Space from 'components/Core/Layout/Space/Space'
+import SectionTitle from 'components/Core/Text/SectionTitle'
+import AccountInformationFieldset from 'components/fieldsets/shared/AccountInformationFieldset'
+import AvailabillityFieldset from 'components/fieldsets/shared/AvailabillityFieldset'
+import ContactPersonInformationFieldset from 'components/fieldsets/shared/ContactPersonInformationFieldset'
+import ContactInformationFieldset from 'components/fieldsets/shared/ContactPersonInformationFieldset'
+import CourseInformationFieldset from 'components/fieldsets/shared/CourseInformationFieldset'
+import EducationInformationFieldset from 'components/fieldsets/shared/EducationInformationFieldset'
+import InformationFieldset from 'components/fieldsets/shared/InformationFieldset'
+import PersonInformationFieldset, { Roles } from 'components/fieldsets/shared/PersonInformationFieldset'
+import { useMockMutation } from 'hooks/UseMockMutation'
+import { routes } from 'routes/routes'
+import { SupplierDetailParams } from 'routes/supplier/types'
+import { Forms } from 'utils/forms'
 import { coworkersCreateMock } from './mocks/coworkers'
 
 interface FormModel {
