+++ client/src/views/Authorized/Supplier/BiscView/SupplierDetailView/Coworkers/CoworkersOverviewView.tsx
@@ -2,27 +2,27 @@ import { t } from '@lingui/macro'
 import { useLingui } from '@lingui/react'
 import React from 'react'
 import { useHistory, useParams } from 'react-router-dom'
-import Headline from '../../../../../components/Chrome/Headline'
-import Breadcrumb from '../../../../../components/Core/Breadcrumb/Breadcrumb'
-import Breadcrumbs from '../../../../../components/Core/Breadcrumb/Breadcrumbs'
-import Button from '../../../../../components/Core/Button/Button'
-import RoleLabelTag from '../../../../../components/Core/DataDisplay/LabelTag/RoleLabelTag'
-import ErrorBlock from '../../../../../components/Core/Feedback/Error/ErrorBlock'
-import Spinner, { Animation } from '../../../../../components/Core/Feedback/Spinner/Spinner'
-import { IconType } from '../../../../../components/Core/Icon/IconType'
-import Center from '../../../../../components/Core/Layout/Center/Center'
-import Column from '../../../../../components/Core/Layout/Column/Column'
-import Row from '../../../../../components/Core/Layout/Row/Row'
-import { Table } from '../../../../../components/Core/Table/Table'
-import { TableLink } from '../../../../../components/Core/Table/TableLink'
-import Tab from '../../../../../components/Core/TabSwitch/Tab'
-import TabSwitch from '../../../../../components/Core/TabSwitch/TabSwitch'
-import { TabProps } from '../../../../../components/Core/TabSwitch/types'
-import { useAanbiederEmployeesQuery } from '../../../../../generated/graphql'
-import { routes } from '../../../../../routes/routes'
-import { SupplierDetailParams } from '../../../../../routes/supplier/types'
-import { DateFormatters } from '../../../../../utils/formatters/Date/Date'
-import { NameFormatters } from '../../../../../utils/formatters/name/Name'
+import Headline from 'components/Chrome/Headline'
+import Breadcrumb from 'components/Core/Breadcrumb/Breadcrumb'
+import Breadcrumbs from 'components/Core/Breadcrumb/Breadcrumbs'
+import Button from 'components/Core/Button/Button'
+import RoleLabelTag from 'components/Core/DataDisplay/LabelTag/RoleLabelTag'
+import ErrorBlock from 'components/Core/Feedback/Error/ErrorBlock'
+import Spinner, { Animation } from 'components/Core/Feedback/Spinner/Spinner'
+import { IconType } from 'components/Core/Icon/IconType'
+import Center from 'components/Core/Layout/Center/Center'
+import Column from 'components/Core/Layout/Column/Column'
+import Row from 'components/Core/Layout/Row/Row'
+import { Table } from 'components/Core/Table/Table'
+import { TableLink } from 'components/Core/Table/TableLink'
+import Tab from 'components/Core/TabSwitch/Tab'
+import TabSwitch from 'components/Core/TabSwitch/TabSwitch'
+import { TabProps } from 'components/Core/TabSwitch/types'
+import { useAanbiederEmployeesQuery } from 'generated/graphql'
+import { routes } from 'routes/routes'
+import { SupplierDetailParams } from 'routes/supplier/types'
+import { DateFormatters } from 'utils/formatters/Date/Date'
+import { NameFormatters } from 'utils/formatters/name/Name'
 
 interface Props {}
