+++ client/src/views/Authorized/Supplier/BiscView/SupplierDetailView/Coworkers/CoworkersView.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Redirect, Route, Switch } from 'react-router-dom'
-import { routes } from '../../../../../routes/routes'
-import { NotFoundView } from '../../../../Generic/NotFoundView'
+import { routes } from 'routes/routes'
+import { NotFoundView } from 'views/Generic/NotFoundView'
 import CoworkersDetailView from './CoworkerDetail/CoworkerDetailView'
 import CoworkerCreateView from './CoworkersCreateView'
 import CoworkersOverviewView from './CoworkersOverviewView'
