+++ client/src/views/Authorized/Supplier/BiscView/SupplierDetailView/Data/DataUpdateView.tsx
@@ -2,32 +2,32 @@ import { t } from '@lingui/macro'
 import { useLingui } from '@lingui/react'
 import React, { useState } from 'react'
 import { useHistory, useParams } from 'react-router-dom'
-import Headline from '../../../../../components/Chrome/Headline'
-import Actionbar from '../../../../../components/Core/Actionbar/Actionbar'
-import Breadcrumb from '../../../../../components/Core/Breadcrumb/Breadcrumb'
-import Breadcrumbs from '../../../../../components/Core/Breadcrumb/Breadcrumbs'
-import Button, { ButtonType } from '../../../../../components/Core/Button/Button'
-import ErrorBlock from '../../../../../components/Core/Feedback/Error/ErrorBlock'
-import { NotificationsManager } from '../../../../../components/Core/Feedback/Notifications/NotificationsManager'
-import Spinner, { Animation } from '../../../../../components/Core/Feedback/Spinner/Spinner'
-import Form from '../../../../../components/Core/Form/Form'
-import HorizontalRule from '../../../../../components/Core/HorizontalRule/HorizontalRule'
-import { IconType } from '../../../../../components/Core/Icon/IconType'
-import Center from '../../../../../components/Core/Layout/Center/Center'
-import Row from '../../../../../components/Core/Layout/Row/Row'
-import Space from '../../../../../components/Core/Layout/Space/Space'
-import Modal from '../../../../../components/Core/Modal/Modal'
+import Headline from 'components/Chrome/Headline'
+import Actionbar from 'components/Core/Actionbar/Actionbar'
+import Breadcrumb from 'components/Core/Breadcrumb/Breadcrumb'
+import Breadcrumbs from 'components/Core/Breadcrumb/Breadcrumbs'
+import Button, { ButtonType } from 'components/Core/Button/Button'
+import ErrorBlock from 'components/Core/Feedback/Error/ErrorBlock'
+import { NotificationsManager } from 'components/Core/Feedback/Notifications/NotificationsManager'
+import Spinner, { Animation } from 'components/Core/Feedback/Spinner/Spinner'
+import Form from 'components/Core/Form/Form'
+import HorizontalRule from 'components/Core/HorizontalRule/HorizontalRule'
+import { IconType } from 'components/Core/Icon/IconType'
+import Center from 'components/Core/Layout/Center/Center'
+import Row from 'components/Core/Layout/Row/Row'
+import Space from 'components/Core/Layout/Space/Space'
+import Modal from 'components/Core/Modal/Modal'
 import BranchInformationFieldset, {
     BranchInformationFieldsetFormModel,
-} from '../../../../../components/fieldsets/shared/BranchInformationFieldset'
+} from 'components/fieldsets/shared/BranchInformationFieldset'
 import ContactInformationFieldset, {
     ContactInformationFieldsetModel,
-} from '../../../../../components/fieldsets/shared/ContactInformationFieldset'
-import { useAanbiederQuery, useUpdateAanbiederMutation } from '../../../../../generated/graphql'
-import { routes } from '../../../../../routes/routes'
-import { SupplierDetailParams } from '../../../../../routes/supplier/types'
-import { Forms } from '../../../../../utils/forms'
-import AanbiederDeleteModalView from '../../Modals/DeleteSupplierModal'
+} from 'components/fieldsets/shared/ContactInformationFieldset'
+import { useAanbiederQuery, useUpdateAanbiederMutation } from 'generated/graphql'
+import { routes } from 'routes/routes'
+import { SupplierDetailParams } from 'routes/supplier/types'
+import { Forms } from 'utils/forms'
+import AanbiederDeleteModalView from 'components/Domain/Bisc/Suppliers/DeleteSupplierModal'
 
 interface FormModel extends BranchInformationFieldsetFormModel, ContactInformationFieldsetModel {}
