+++ client/src/views/Authorized/Supplier/BiscView/SupplierDetailView/Data/DataView.tsx
@@ -2,26 +2,26 @@ import { t } from '@lingui/macro'
 import { useLingui } from '@lingui/react'
 import React from 'react'
 import { useHistory, useParams } from 'react-router-dom'
-import Headline, { SpacingType } from '../../../../../components/Chrome/Headline'
-import Actionbar from '../../../../../components/Core/Actionbar/Actionbar'
-import Breadcrumb from '../../../../../components/Core/Breadcrumb/Breadcrumb'
-import Breadcrumbs from '../../../../../components/Core/Breadcrumb/Breadcrumbs'
-import Button, { ButtonType } from '../../../../../components/Core/Button/Button'
-import ErrorBlock from '../../../../../components/Core/Feedback/Error/ErrorBlock'
-import Spinner, { Animation } from '../../../../../components/Core/Feedback/Spinner/Spinner'
-import HorizontalRule from '../../../../../components/Core/HorizontalRule/HorizontalRule'
-import Center from '../../../../../components/Core/Layout/Center/Center'
-import Column from '../../../../../components/Core/Layout/Column/Column'
-import Row from '../../../../../components/Core/Layout/Row/Row'
-import Space from '../../../../../components/Core/Layout/Space/Space'
-import Tab from '../../../../../components/Core/TabSwitch/Tab'
-import TabSwitch from '../../../../../components/Core/TabSwitch/TabSwitch'
-import { TabProps } from '../../../../../components/Core/TabSwitch/types'
-import BranchInformationFieldset from '../../../../../components/fieldsets/shared/BranchInformationFieldset'
-import ContactInformationFieldset from '../../../../../components/fieldsets/shared/ContactInformationFieldset'
-import { useAanbiederQuery } from '../../../../../generated/graphql'
-import { routes } from '../../../../../routes/routes'
-import { SupplierDetailParams } from '../../../../../routes/supplier/types'
+import Headline, { SpacingType } from 'components/Chrome/Headline'
+import Actionbar from 'components/Core/Actionbar/Actionbar'
+import Breadcrumb from 'components/Core/Breadcrumb/Breadcrumb'
+import Breadcrumbs from 'components/Core/Breadcrumb/Breadcrumbs'
+import Button, { ButtonType } from 'components/Core/Button/Button'
+import ErrorBlock from 'components/Core/Feedback/Error/ErrorBlock'
+import Spinner, { Animation } from 'components/Core/Feedback/Spinner/Spinner'
+import HorizontalRule from 'components/Core/HorizontalRule/HorizontalRule'
+import Center from 'components/Core/Layout/Center/Center'
+import Column from 'components/Core/Layout/Column/Column'
+import Row from 'components/Core/Layout/Row/Row'
+import Space from 'components/Core/Layout/Space/Space'
+import Tab from 'components/Core/TabSwitch/Tab'
+import TabSwitch from 'components/Core/TabSwitch/TabSwitch'
+import { TabProps } from 'components/Core/TabSwitch/types'
+import BranchInformationFieldset from 'components/fieldsets/shared/BranchInformationFieldset'
+import ContactInformationFieldset from 'components/fieldsets/shared/ContactInformationFieldset'
+import { useAanbiederQuery } from 'generated/graphql'
+import { routes } from 'routes/routes'
+import { SupplierDetailParams } from 'routes/supplier/types'
 
 interface Props {}
