+++ client/src/views/Authorized/Supplier/BiscView/SupplierDetailView/SupplierDetailView.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { Redirect, Route, Switch } from 'react-router-dom'
-import { routes } from '../../../../routes/routes'
+import { routes } from 'routes/routes'
 
-import { NotFoundView } from '../../../Generic/NotFoundView'
+import { NotFoundView } from 'views/Generic/NotFoundView'
 import CoworkersView from './Coworkers/CoworkersView'
 import DataUpdateView from './Data/DataUpdateView'
 import DataView from './Data/DataView'
