+++ client/src/views/Authorized/Supplier/BiscView/SupplierOverviewView.tsx
@@ -1,20 +1,20 @@
 import { t } from '@lingui/macro'
 import { useLingui } from '@lingui/react'
 import React from 'react'
-import { Link, useHistory } from 'react-router-dom'
-import Headline, { SpacingType } from '../../../components/Chrome/Headline'
-import Button from '../../../components/Core/Button/Button'
-import ErrorBlock from '../../../components/Core/Feedback/Error/ErrorBlock'
-import Spinner, { Animation } from '../../../components/Core/Feedback/Spinner/Spinner'
-import { IconType } from '../../../components/Core/Icon/IconType'
-import Center from '../../../components/Core/Layout/Center/Center'
-import Column from '../../../components/Core/Layout/Column/Column'
-import Row from '../../../components/Core/Layout/Row/Row'
-import { Table } from '../../../components/Core/Table/Table'
-import { TableLink } from '../../../components/Core/Table/TableLink'
-import { useAanbiedersQuery } from '../../../generated/graphql'
-import { routes } from '../../../routes/routes'
-import { AdressFormatters } from '../../../utils/formatters/Address/Address'
+import { useHistory } from 'react-router-dom'
+import Headline, { SpacingType } from 'components/Chrome/Headline'
+import Button from 'components/Core/Button/Button'
+import ErrorBlock from 'components/Core/Feedback/Error/ErrorBlock'
+import Spinner, { Animation } from 'components/Core/Feedback/Spinner/Spinner'
+import { IconType } from 'components/Core/Icon/IconType'
+import Center from 'components/Core/Layout/Center/Center'
+import Column from 'components/Core/Layout/Column/Column'
+import Row from 'components/Core/Layout/Row/Row'
+import { Table } from 'components/Core/Table/Table'
+import { TableLink } from 'components/Core/Table/TableLink'
+import { useAanbiedersQuery } from 'generated/graphql'
+import { routes } from 'routes/routes'
+import { AdressFormatters } from 'utils/formatters/Address/Address'
 
 interface Props {}
